Rename CellChangesConfirmation.tsx to CellChangesContainer.tsx
